+++ derive/src/args.rs
@@ -338,6 +338,8 @@ pub struct Enum {
     #[darling(default)]
     pub name: Option<String>,
     #[darling(default)]
+    pub name_type: bool,
+    #[darling(default)]
     pub rename_items: Option<RenameRule>,
     #[darling(default)]
     pub remote: Option<String>,
@@ -380,6 +382,8 @@ pub struct Union {
     pub internal: bool,
     #[darling(default)]
     pub name: Option<String>,
+    #[darling(default)]
+    pub name_type: bool,
     #[darling(default)]
     pub visible: Option<Visible>,
     #[darling(default)]
@@ -448,6 +452,8 @@ pub struct InputObject {
     #[darling(default)]
     pub name: Option<String>,
     #[darling(default)]
+    pub name_type: bool,
+    #[darling(default)]
     pub input_name: Option<String>,
     #[darling(default)]
     pub rename_fields: Option<RenameRule>,
@@ -459,6 +465,8 @@ pub struct InputObject {
     pub tags: Vec<String>,
     #[darling(default, multiple, rename = "concrete")]
     pub concretes: Vec<ConcreteType>,
+    #[darling(default)]
+    pub validator: Option<SpannedValue<String>>,
     // for SimpleObject
     #[darling(default)]
     pub complex: bool,
@@ -497,6 +505,8 @@ pub struct OneofObject {
     pub internal: bool,
     #[darling(default)]
     pub name: Option<String>,
+    #[darling(default)]
+    pub name_type: bool,
     #[darling(default)]
     pub rename_fields: Option<RenameRule>,
     #[darling(default)]
@@ -580,6 +590,8 @@ pub struct Interface {
     #[darling(default)]
     pub name: Option<String>,
     #[darling(default)]
+    pub name_type: bool,
+    #[darling(default)]
     pub rename_fields: Option<RenameRule>,
     #[darling(default)]
     pub rename_args: Option<RenameRule>,
@@ -600,6 +612,8 @@ pub struct Interface {
 pub struct Scalar {
     pub internal: bool,
     pub name: Option<String>,
+    #[darling(default)]
+    pub name_type: bool,
     pub use_type_description: bool,
     pub visible: Option<Visible>,
     pub inaccessible: bool,
@@ -613,6 +627,8 @@ pub struct Scalar {
 pub struct Subscription {
     pub internal: bool,
     pub name: Option<String>,
+    #[darling(default)]
+    pub name_type: bool,
     pub rename_fields: Option<RenameRule>,
     pub rename_args: Option<RenameRule>,
     pub use_type_description: bool,
@@ -665,6 +681,8 @@ pub struct MergedObject {
     #[darling(default)]
     pub name: Option<String>,
     #[darling(default)]
+    pub name_type: bool,
+    #[darling(default)]
     pub cache_control: CacheControl,
     #[darling(default)]
     pub extends: bool,
@@ -697,6 +715,8 @@ pub struct MergedSubscription {
     pub internal: bool,
     #[darling(default)]
     pub name: Option<String>,
+    #[darling(default)]
+    pub name_type: bool,
     #[darling(default)]
     pub visible: Option<Visible>,
     #[darling(default)]
@@ -833,7 +853,6 @@ pub struct NewType {
 #[darling(default)]
 pub struct ComplexObject {
     pub internal: bool,
-    pub name: Option<String>,
     pub rename_fields: Option<RenameRule>,
     pub rename_args: Option<RenameRule>,
     pub guard: Option<SpannedValue<String>>,
@@ -867,6 +886,8 @@ pub struct ComplexObjectField {
 pub struct Directive {
     pub internal: bool,
     pub name: Option<String>,
+    #[darling(default)]
+    pub name_type: bool,
     pub visible: Option<Visible>,
     pub repeatable: bool,
     pub rename_args: Option<RenameRule>,